Refactor: create DB module
