+++ main.py
@@ -1,12 +1,9 @@
-import os
 from typing import Annotated, Optional
-import dotenv
 import datetime
-from asyncpg import Pool, create_pool as asyncpg_create_pool
-from fastapi import FastAPI, Depends, HTTPException, Query, Request
-from fastapi.responses import JSONResponse
+from fastapi import FastAPI, HTTPException, Query
+
+# from fastapi.responses import JSONResponse
 from sqlmodel import (
-    Session,
     SQLModel,
     create_engine,
     Column,
@@ -18,6 +15,7 @@ from sqlmodel import (
 )
 from pydantic import BaseModel
 from auth.router import auth_router
+from db import get_session, run_query, init_postgres
 
 
 # Project metadata models
@@ -119,110 +117,11 @@ class ParticulateMatterData(SensorData):
     value: dict = Field(sa_column=Column(JSON), default={})
 
 
-sensor_data_hypertables = {
-    "PM_data": "sensor_PM_data",
-    "temp_humidity": "sensor_temp_humidity_data",
-}
-
-dotenv.load_dotenv(override=True)
-TIMESCALE_DB_CONNECTION = os.getenv("TIMESCALE_DB_CONNECTION")
-print(TIMESCALE_DB_CONNECTION)
-
 sqlite_file_name = "sensorsafrica.db"
 sqlite_url = f"sqlite:///{sqlite_file_name}"
 
 connect_args = {"check_same_thread": False}
 engine = create_engine(sqlite_url, connect_args=connect_args, echo=True)
-postgres_engine = create_engine(TIMESCALE_DB_CONNECTION, echo=True)
-
-
-tsb_conn_pool: Optional[Pool] = None
-
-# Sensor Data Table
-create_hypertable_query = f"""
-
-DROP TABLE IF EXISTS {sensor_data_hypertables["PM_data"]};
-DROP TABLE IF EXISTS {sensor_data_hypertables["temp_humidity"]};
-
-CREATE TABLE IF NOT EXISTS {sensor_data_hypertables["PM_data"]} (
-    time TIMESTAMPTZ NOT NULL,
-    node_id VARCHAR(30) NOT NULL,
-    PM1 FLOAT,
-    PM2_5 FLOAT,
-    PM10 FLOAT,
-    location VARCHAR(64) NOT NULL,
-    sensor_name VARCHAR(64) NOT NULL,
-    FOREIGN KEY (node_id) REFERENCES node(node_id)
-);
-CREATE TABLE IF NOT EXISTS {sensor_data_hypertables["temp_humidity"]} (
-    time TIMESTAMPTZ NOT NULL,
-    node_id VARCHAR(30) NOT NULL,
-    temperature FLOAT,
-    rel_hum FLOAT,
-    abs_hum FLOAT,
-    heat_index FLOAT,
-    location VARCHAR(64) NOT NULL,
-    sensor_name VARCHAR(64) NOT NULL,
-    FOREIGN KEY (node_id) REFERENCES node(node_id)
-);
-
-SELECT create_hypertable('{sensor_data_hypertables["PM_data"]}', 'time', if_not_exists => TRUE);
-SELECT create_hypertable('{sensor_data_hypertables["temp_humidity"]}', 'time', if_not_exists => TRUE);
-"""
-
-
-async def init_connection_pool():
-    global tsb_conn_pool
-    try:
-        print("Initializing PostgreSQL connection pool...")
-        tsb_conn_pool = await asyncpg_create_pool(
-            dsn=TIMESCALE_DB_CONNECTION, min_size=1, max_size=10
-        )
-        print("PostgreSQL connection pool created successfully.")
-
-    except Exception as e:
-        print(f"Error initializing PostgreSQL connection pool: {e}")
-        raise
-
-
-async def run_query(query):
-    global tsb_conn_pool
-    try:
-        conn = await tsb_conn_pool.execute(query)
-        return conn
-    except Exception as e:
-        print(f"Error occured when running query : {e}")
-        print(query)
-        raise
-
-
-async def init_postgres() -> None:
-    """
-    Initialize the PostgreSQL connection pool and create the hypertables.
-    """
-    await init_connection_pool()
-
-    # creat tables
-    create_db_and_tables()
-
-    # create hypertables
-    await run_query(create_hypertable_query)
-
-    # # insert dummy data
-    # await run_query(dummy_sensor_data_query)
-
-
-def create_db_and_tables():
-    SQLModel.metadata.create_all(engine)
-    SQLModel.metadata.create_all(postgres_engine)
-
-
-def get_session():
-    with Session(postgres_engine) as session:
-        yield session
-
-
-SessionDep = Annotated[Session, Depends(get_session)]
 
 
 app = FastAPI()
